move ws to api/permissions
